app/requirements.txt
streamlit
pandas
numpy
scikit-learn
requests
plotly
google-generativeai
matplotlib
statsmodels
